refactor: move shared chart components and unify supabase client (Phase 1-3) #86ex8bhzz
Phase 1: Move shared chart components to components/charts/
- chart-card, new-chart-button, completed-charts-section, delete-chart-button
- Original files now re-export from new location for backward compatibility

Phase 2: Move app/charts/actions.ts to lib/charts-actions.ts
- createChart, getChartsHierarchy, deleteChart, archiveChart, restoreChart
- Original file now re-exports from new location
- Updated 9 direct importers to use @/lib/charts-actions

Phase 3: Unify Supabase client imports
- @/utils/supabase/server → @/lib/supabase/server (7 files)
- Deleted utils/supabase/ directory (server.ts + middleware.ts, both unused)
