+++ app/charts/[id]/actions.ts
@@ -2,7 +2,7 @@
 
 import { revalidatePath } from "next/cache";
 import { supabase } from "@/lib/supabase";
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@/lib/supabase/server";
 
 async function revalidateChartPath(chartId: string) {
   revalidatePath(`/charts/${chartId}`);
