+++ app/workspaces/[wsId]/charts/[id]/actions.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { revalidatePath } from "next/cache";
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@/lib/supabase/server";
 
 async function revalidateChartPath(chartId: string) {
   const supabaseClient = await createClient();
