+++ app/workspaces/[wsId]/charts/[id]/kanban/page.tsx
@@ -1,5 +1,5 @@
 export const dynamic = "force-dynamic";
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@/lib/supabase/server";
 import { notFound } from "next/navigation";
 import { getWorkspaceMembers } from "@/lib/workspace";
 import { KanbanBoard } from "./kanban-board";
