+++ components/chart-switcher.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react";
 import { useParams, usePathname } from "next/navigation";
 import { useLocale } from "next-intl";
-import { getChartsHierarchy, type ProjectGroup } from "@/app/charts/actions";
+import { getChartsHierarchy, type ProjectGroup } from "@/lib/charts-actions";
 import { ChevronDown, BarChart3, ExternalLink } from "lucide-react";
 import { supabase } from "@/lib/supabase";
 import {
